Create avengersquizcode.py
This file contains the rough draft of the code that I will be using for the project. This code was forked to my profile into my own repository, which I now copy and pasted into this file. This code is from Filla0 and his repository called Which Avenger are You.

Link: https://github.com/ckinsing/Which-Avenger-are-you 

I will be updating and personalizing this code soon to fit the contents of my quiz ideas.